Moved tests to tests/
